fix(frontend): exclude type-guards from vitest + remove unused eslint-disable
