+++ frontend/components/layout/ResultPanel.tsx
@@ -4,10 +4,13 @@ import { useState, useMemo } from "react";
 import dynamic from "next/dynamic";
 import type { RuntimeResponse, PilgrimagePoint, SearchResultData } from "../../lib/types";
 import { isSearchData } from "../../lib/types";
-import { useDict } from "../../lib/i18n-context";
 import { usePointSelectionContext } from "../../contexts/PointSelectionContext";
+import { useDict } from "../../lib/i18n-context";
 import SelectionBar from "../generative/SelectionBar";
 import GenerativeUIRenderer from "../generative/GenerativeUIRenderer";
+import { PhotoCard } from "../generative/PhotoCard";
+import { ResultPanelToolbar } from "./ResultPanelToolbar";
+import { ResultPanelEmptyState } from "./ResultPanelEmptyState";
 
 // ---------------------------------------------------------------------------
 // Leaflet map — lazy-loaded with ssr:false so Leaflet's window accesses do not
@@ -62,78 +65,6 @@ function buildEpRanges(points: PilgrimagePoint[]): string[] {
 }
 
 // ---------------------------------------------------------------------------
-// Sub-components
-// ---------------------------------------------------------------------------
-
-function EpisodeBadge({ episode }: { episode: number | null }) {
-  const { grid: t } = useDict();
-  if (episode == null) return null;
-  return (
-    <span
-      className="absolute left-2 top-2 rounded-[5px] px-2 py-0.5 text-[10px] font-semibold tracking-wide text-white"
-      style={{ background: "rgba(0,0,0,0.55)", backdropFilter: "blur(4px)" }}
-    >
-      {t.episode.replace("{ep}", String(episode))}
-    </span>
-  );
-}
-
-interface PhotoCardProps {
-  point: PilgrimagePoint;
-  selected: boolean;
-  onToggle: (id: string) => void;
-}
-
-function PhotoCard({ point, selected, onToggle }: PhotoCardProps) {
-  return (
-    <div
-      className="group relative cursor-pointer overflow-hidden rounded-[10px] bg-[var(--color-card)] transition-all duration-200 hover:-translate-y-0.5 hover:shadow-lg"
-      style={{ border: `2px solid ${selected ? "var(--color-primary)" : "transparent"}` }}
-      onClick={() => onToggle(point.id)}
-      role="button"
-      tabIndex={0}
-      aria-pressed={selected}
-      onKeyDown={(e) => {
-        if (e.key === "Enter" || e.key === " ") {
-          e.preventDefault();
-          onToggle(point.id);
-        }
-      }}
-    >
-      {/* Image — 16/10 aspect ratio */}
-      <div className="relative overflow-hidden" style={{ aspectRatio: "16/10" }}>
-        {point.screenshot_url ? (
-          <img
-            src={point.screenshot_url}
-            alt={point.name}
-            className="h-full w-full object-cover transition-transform duration-300 group-hover:scale-[1.04]"
-          />
-        ) : (
-          <div className="h-full w-full bg-[var(--color-muted)]" />
-        )}
-        <EpisodeBadge episode={point.episode} />
-        {selected && (
-          <div className="absolute right-2 top-2 flex h-[22px] w-[22px] items-center justify-center rounded-full bg-[var(--color-primary)] text-[11px] font-bold text-white">
-            ✓
-          </div>
-        )}
-      </div>
-      {/* Info */}
-      <div className="px-3 py-2.5">
-        <p className="truncate text-[13px] font-medium text-[var(--color-fg)]">
-          {point.name}
-        </p>
-        {point.title && (
-          <p className="mt-0.5 truncate text-[11px] text-[var(--color-muted-fg)]">
-            {point.title}
-          </p>
-        )}
-      </div>
-    </div>
-  );
-}
-
-// ---------------------------------------------------------------------------
 // Loading skeleton
 // ---------------------------------------------------------------------------
 
@@ -154,193 +85,6 @@ function LoadingSkeleton() {
         className="mt-2 h-32 w-full rounded-sm bg-[var(--color-muted)]"
         style={{ animation: "pulse-skeleton 1.6s ease-in-out infinite 0.2s" }}
       />
-    </div>
-  );
-}
-
-// ---------------------------------------------------------------------------
-// Empty state
-// ---------------------------------------------------------------------------
-
-function EmptyState({
-  onSuggest,
-}: {
-  onSuggest?: (text: string) => void;
-}) {
-  const { chat, clarification } = useDict();
-
-  return (
-    <div className="relative flex min-h-0 flex-1 overflow-hidden">
-      {/* Radial gradient background */}
-      <div
-        className="pointer-events-none absolute inset-0"
-        style={{
-          background:
-            "radial-gradient(ellipse 70% 60% at 60% 70%, oklch(93% 0.025 240 / 0.18), transparent 70%)",
-        }}
-      />
-      {/* Map watermark */}
-      <div
-        className="pointer-events-none absolute inset-0 opacity-[0.05]"
-        style={{
-          backgroundImage: "url(/empty-map.svg)",
-          backgroundSize: "cover",
-          backgroundPosition: "center",
-        }}
-      />
-
-      <div className="relative flex flex-1 flex-col justify-end pb-16 pl-10 pr-8">
-        <div className="relative mb-6 select-none leading-[0.85]">
-          <div
-            className="font-[family-name:var(--app-font-display)] font-bold"
-            style={{
-              fontSize: "clamp(5rem, 12vw, 9rem)",
-              color: "color-mix(in oklch, var(--color-fg) 9%, transparent)",
-            }}
-          >
-            <div>聖地</div>
-            <div>巡礼</div>
-          </div>
-          <div
-            className="absolute left-0 top-0 font-[family-name:var(--app-font-display)] font-bold text-[var(--color-primary)]"
-            style={{ fontSize: "clamp(5rem, 12vw, 9rem)", lineHeight: "0.85" }}
-          >
-            聖
-          </div>
-        </div>
-
-        <div className="mb-5 w-12 border-t border-[var(--color-border)]" />
-
-        <p className="mb-8 max-w-xs text-sm font-light leading-relaxed text-[var(--color-muted-fg)]">
-          {chat.welcome_subtitle}
-        </p>
-
-        <div className="flex flex-col gap-1.5">
-          {clarification.suggestions.map((s) => (
-            <button
-              key={s.label}
-              onClick={() => onSuggest?.(s.query)}
-              className="w-fit text-left text-xs font-light text-[var(--color-muted-fg)] transition-colors hover:text-[var(--color-primary)]"
-              style={{ transitionDuration: "var(--duration-fast)" }}
-            >
-              {s.label} →
-            </button>
-          ))}
-        </div>
-      </div>
-    </div>
-  );
-}
-
-// ---------------------------------------------------------------------------
-// Toolbar: view toggle + filter chips
-// ---------------------------------------------------------------------------
-
-interface ToolbarProps {
-  view: ViewMode;
-  onViewChange: (v: ViewMode) => void;
-  epRanges: string[];
-  activeEpRange: string | null;
-  onEpRangeChange: (range: string | null) => void;
-}
-
-function Toolbar({
-  view,
-  onViewChange,
-  epRanges,
-  activeEpRange,
-  onEpRangeChange,
-}: ToolbarProps) {
-  return (
-    <div className="flex shrink-0 items-center gap-2 border-b border-[var(--color-border)] bg-white px-4 py-1.5">
-      {/* Episode filter chips */}
-      {epRanges.length > 0 && (
-        <div className="flex items-center gap-1.5 overflow-x-auto">
-          <button
-            onClick={() => onEpRangeChange(null)}
-            className="shrink-0 rounded-[18px] border px-3 py-1 text-[11px] font-medium transition-all duration-150"
-            style={
-              activeEpRange === null
-                ? {
-                    background: "var(--color-primary)",
-                    color: "white",
-                    borderColor: "var(--color-primary)",
-                  }
-                : {
-                    background: "white",
-                    color: "var(--color-muted-fg)",
-                    borderColor: "var(--color-border)",
-                  }
-            }
-          >
-            すべて
-          </button>
-          {epRanges.map((range) => (
-            <button
-              key={range}
-              onClick={() => onEpRangeChange(range)}
-              className="shrink-0 rounded-[18px] border px-3 py-1 text-[11px] font-medium transition-all duration-150"
-              style={
-                activeEpRange === range
-                  ? {
-                      background: "var(--color-primary)",
-                      color: "white",
-                      borderColor: "var(--color-primary)",
-                    }
-                  : {
-                      background: "white",
-                      color: "var(--color-muted-fg)",
-                      borderColor: "var(--color-border)",
-                    }
-              }
-            >
-              {range}
-            </button>
-          ))}
-        </div>
-      )}
-
-      {/* Spacer */}
-      <div className="flex-1" />
-
-      {/* Grid / map pill toggle */}
-      <div
-        className="flex shrink-0 gap-0.5 rounded-lg p-0.5"
-        style={{ background: "var(--color-card)" }}
-      >
-        <button
-          onClick={() => onViewChange("grid")}
-          className="flex items-center gap-1.5 rounded-md px-3.5 py-1.5 text-[12px] font-medium transition-all duration-150"
-          style={
-            view === "grid"
-              ? {
-                  background: "white",
-                  color: "var(--color-fg)",
-                  boxShadow: "0 1px 3px rgba(0,0,0,0.08)",
-                }
-              : { background: "transparent", color: "var(--color-muted-fg)" }
-          }
-        >
-          <span>📷</span>
-          グリッド
-        </button>
-        <button
-          onClick={() => onViewChange("map")}
-          className="flex items-center gap-1.5 rounded-md px-3.5 py-1.5 text-[12px] font-medium transition-all duration-150"
-          style={
-            view === "map"
-              ? {
-                  background: "white",
-                  color: "var(--color-fg)",
-                  boxShadow: "0 1px 3px rgba(0,0,0,0.08)",
-                }
-              : { background: "transparent", color: "var(--color-muted-fg)" }
-          }
-        >
-          <span>🗺</span>
-          マップ
-        </button>
-      </div>
     </div>
   );
 }
@@ -449,7 +193,7 @@ export default function ResultPanel({
     return (
       <section className="relative flex min-h-0 min-w-0 flex-1 flex-col overflow-hidden bg-[var(--color-bg)]">
         {selectionBar}
-        <EmptyState onSuggest={onSuggest} />
+        <ResultPanelEmptyState onSuggest={onSuggest} />
       </section>
     );
   }
@@ -466,7 +210,7 @@ export default function ResultPanel({
         {selectionBar}
 
         {/* Toolbar: filter chips + view toggle */}
-        <Toolbar
+        <ResultPanelToolbar
           view={view}
           onViewChange={setView}
           epRanges={epRanges}
